Medical_Bot (1)/requirements.txt
flask
openai
sentence-transformers
faiss-cpu
chroma
python-dotenv
numpy
flask>=3.0